Move tests to upstream crates
Test cases for Outpoint, Version, CompactTarget and Sequence have been
left in bitcoin even following their moves to primitives and units. In
order to keep tests close to the corresponding types, the tests should
be moved to the upstream leaf crates.

Move tests to units and primitives alongside their corresponding types.
