+++ bitcoin/src/pow.rs
@@ -662,19 +662,6 @@ mod tests {
     }
 
     #[test]
-    fn roundtrip_compact_target() {
-        let consensus = 0x1d00_ffff;
-        let compact = CompactTarget::from_consensus(consensus);
-        let t = Target::from_compact(CompactTarget::from_consensus(consensus));
-        assert_eq!(t, Target::from(compact)); // From/Into sanity check.
-
-        let back = t.to_compact_lossy();
-        assert_eq!(back, compact); // From/Into sanity check.
-
-        assert_eq!(back.to_consensus(), consensus);
-    }
-
-    #[test]
     fn roundtrip_target_work() {
         let target = u32_to_target(0xdeadbeef_u32);
         let work = target.to_work();
