+++ primitives/src/block.rs
@@ -1059,6 +1059,24 @@ mod tests {
     }
 
     #[test]
+    fn soft_fork_signalling() {
+        for i in 0..31 {
+            let version_int = (0x20000000u32 ^ (1 << i)) as i32;
+            let version = Version::from_consensus(version_int);
+            if i < 29 {
+                assert!(version.is_signalling_soft_fork(i));
+            } else {
+                assert!(!version.is_signalling_soft_fork(i));
+            }
+        }
+
+        let segwit_signal = Version::from_consensus(0x20000000 ^ (1 << 1));
+        assert!(!segwit_signal.is_signalling_soft_fork(0));
+        assert!(segwit_signal.is_signalling_soft_fork(1));
+        assert!(!segwit_signal.is_signalling_soft_fork(2));
+    }
+
+    #[test]
     fn version_to_consensus() {
         let version = Version::from_consensus(1_234_567_890);
         assert_eq!(version.to_consensus(), 1_234_567_890);
